[VS-7337] restitution API Startup removed, Program cleanup to minimize amount of dead code
