+++ restitution-app/Program.cs
@@ -1,48 +1,332 @@
-using Microsoft.AspNetCore;
-using Microsoft.AspNetCore.Hosting;
+using System;
+using System.IO;
+using System.Net.Http;
+using System.Reflection;
+using Database.Extensions;
+using Gov.Cscp.VictimServices.Public.Services;
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.CookiePolicy;
+using Microsoft.AspNetCore.DataProtection;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Http.Features;
 using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Logging;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Diagnostics.HealthChecks;
+using Microsoft.Extensions.Hosting;
+using NWebsec.AspNetCore.Mvc;
+using NWebsec.AspNetCore.Mvc.Csp;
+using Serilog;
+using Serilog.Enrichers.Span;
+using Serilog.Exceptions;
 
-namespace Gov.Cscp.VictimServices.Public
-{
-    public class Program
+var builder = WebApplication.CreateBuilder(args);
+
+// ---------------------------------------------------------------------------
+// Configuration
+// ---------------------------------------------------------------------------
+builder
+    .Configuration.AddJsonFile("appsettings.json", optional: true, reloadOnChange: true)
+    .AddJsonFile(
+        $"appsettings.{builder.Environment.EnvironmentName}.json",
+        optional: true,
+        reloadOnChange: true
+    )
+    .AddEnvironmentVariables();
+
+// ---------------------------------------------------------------------------
+// Logging (Serilog)
+// ---------------------------------------------------------------------------
+ConfigureLogging(builder.Environment, builder.Configuration);
+builder.Host.UseSerilog();
+
+// ---------------------------------------------------------------------------
+// Services
+// ---------------------------------------------------------------------------
+var services = builder.Services;
+
+// Dataverse / Database
+services.AddDatabase(builder.Configuration);
+
+// Application services
+services.AddScoped<ILookupQueryService, LookupQueryService>();
+
+// Routing
+services.AddRouting(options => options.LowercaseUrls = true);
+
+// MVC + security headers (NWebsec filters)
+services
+    .AddControllers(opts =>
     {
-        public static void Main(string[] args)
+        opts.Filters.Add(typeof(NoCacheHttpHeadersAttribute));
+        opts.Filters.Add(new XRobotsTagAttribute() { NoIndex = true, NoFollow = true });
+        opts.Filters.Add(typeof(XContentTypeOptionsAttribute));
+        opts.Filters.Add(typeof(XDownloadOptionsAttribute));
+        opts.Filters.Add(typeof(XFrameOptionsAttribute));
+        opts.Filters.Add(typeof(XXssProtectionAttribute));
+        opts.Filters.Add(typeof(CspReportOnlyAttribute));
+        opts.Filters.Add(new CspScriptSrcReportOnlyAttribute { None = true });
+    })
+    .AddNewtonsoftJson(opts =>
+    {
+        opts.SerializerSettings.Formatting = Newtonsoft.Json.Formatting.Indented;
+        opts.SerializerSettings.DateFormatHandling = Newtonsoft
+            .Json
+            .DateFormatHandling
+            .IsoDateFormat;
+        opts.SerializerSettings.DateTimeZoneHandling = Newtonsoft.Json.DateTimeZoneHandling.Utc;
+
+        // Prevent JSON parser issues with circular references in user / roles model.
+        opts.SerializerSettings.ReferenceLoopHandling = Newtonsoft
+            .Json
+            .ReferenceLoopHandling
+            .Ignore;
+    });
+
+// Data Protection key ring persistence (for container deployments)
+if (!string.IsNullOrEmpty(builder.Configuration["KEY_RING_DIRECTORY"]))
+{
+    services
+        .AddDataProtection()
+        .PersistKeysToFileSystem(new DirectoryInfo(builder.Configuration["KEY_RING_DIRECTORY"]!));
+}
+
+// Allow large file uploads (1 GB)
+services.Configure<FormOptions>(options =>
+{
+    options.MultipartBodyLengthLimit = 1073741824;
+});
+
+// Health checks
+services.AddHealthChecks().AddCheck("HTTP Endpoint", () => HealthCheckResult.Healthy("Ok"));
+
+// Swagger / OpenAPI
+services.AddSwaggerGen(c =>
+{
+    c.SwaggerDoc(
+        "v1",
+        new Microsoft.OpenApi.OpenApiInfo
         {
-            var host = CreateWebHostBuilder(args).Build();
-            host.Run();
+            Title = "Restitution API",
+            Version = "v1",
+            Description = "API for the Restitution Application",
+        }
+    );
+});
+
+// ---------------------------------------------------------------------------
+// Build
+// ---------------------------------------------------------------------------
+var app = builder.Build();
+
+// ---------------------------------------------------------------------------
+// Middleware pipeline
+// ---------------------------------------------------------------------------
+string pathBase = builder.Configuration["BASE_PATH"];
+if (!string.IsNullOrEmpty(pathBase))
+{
+    app.UsePathBase(pathBase);
+}
+
+if (!app.Environment.IsProduction())
+{
+    app.UseDeveloperExceptionPage();
+}
+else
+{
+    app.UseExceptionHandler("/Home/Error");
+}
+
+app.UseSerilogRequestLogging(options =>
+{
+    options.GetLevel = (httpContext, elapsed, ex) =>
+    {
+        if (ex != null)
+            return Serilog.Events.LogEventLevel.Error;
+
+        var path = httpContext.Request.Path.ToString();
+
+        var logIgnoreEndpoints = new[] { "/hc", "/api/lookup" };
+
+        if (
+            Array.Exists(
+                logIgnoreEndpoints,
+                e => path.StartsWith(e, StringComparison.OrdinalIgnoreCase)
+            )
+        )
+        {
+            return Serilog.Events.LogEventLevel.Verbose;
         }
 
-        public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-            WebHost
-                .CreateDefaultBuilder(args)
-                .ConfigureAppConfiguration(
-                    (hostingContext, config) =>
-                    {
-                        var env = hostingContext.HostingEnvironment;
+        return elapsed > 1000
+            ? Serilog.Events.LogEventLevel.Warning
+            : Serilog.Events.LogEventLevel.Information;
+    };
 
-                        config
-                            .AddJsonFile("appsettings.json", optional: true, reloadOnChange: true)
-                            .AddJsonFile(
-                                $"appsettings.{env.EnvironmentName}.json",
-                                optional: true,
-                                reloadOnChange: true
-                            );
-                        config.AddEnvironmentVariables();
-                    }
-                )
-                .ConfigureLogging(
-                    (hostingContext, logging) =>
-                    {
-                        logging.AddSimpleConsole(opts =>
-                        {
-                            opts.IncludeScopes = true;
-                        });
-                        logging.SetMinimumLevel(LogLevel.Debug);
-                        logging.AddDebug();
-                        logging.AddEventSourceLogger();
-                    }
-                )
-                .UseStartup<Startup>();
+    options.EnrichDiagnosticContext = (diagnosticContext, httpContext) =>
+    {
+        diagnosticContext.Set("RequestHost", httpContext.Request.Host.Value);
+        diagnosticContext.Set("RequestScheme", httpContext.Request.Scheme);
+        diagnosticContext.Set("UserAgent", httpContext.Request.Headers["User-Agent"].ToString());
+    };
+});
+
+// Health checks
+app.UseHealthChecks("/hc");
+
+// Swagger (development only)
+if (app.Environment.IsDevelopment())
+{
+    app.UseSwagger();
+    app.UseSwaggerUI(c =>
+    {
+        c.SwaggerEndpoint("/swagger/v1/swagger.json", "Restitution API v1");
+        c.RoutePrefix = "swagger";
+    });
+}
+
+// Security headers
+app.Use(
+    async (ctx, next) =>
+    {
+        ctx.Response.Headers.Append(
+            "Content-Security-Policy",
+            "script-src 'self' 'unsafe-eval' 'unsafe-inline' https://apis.google.com https://maxcdn.bootstrapcdn.com https://cdnjs.cloudflare.com https://code.jquery.com https://stackpath.bootstrapcdn.com https://fonts.googleapis.com"
+        );
+        ctx.Response.Headers.Append(
+            "Strict-Transport-Security",
+            "max-age=31536000; includeSubDomains; preload"
+        );
+        await next();
     }
+);
+
+app.UseXContentTypeOptions();
+app.UseReferrerPolicy(opts => opts.NoReferrer());
+app.UseXfo(xfo => xfo.Deny());
+
+if (!app.Environment.IsDevelopment())
+{
+    app.UseCsp(opts =>
+    {
+        opts.BlockAllMixedContent()
+            .StyleSources(s =>
+                s.Self()
+                    .UnsafeInline()
+                    .CustomSources(
+                        "https://use.fontawesome.com",
+                        "https://stackpath.bootstrapcdn.com",
+                        "https://fonts.googleapis.com"
+                    )
+            )
+            .FontSources(s =>
+                s.Self().CustomSources("https://use.fontawesome.com", "https://fonts.gstatic.com")
+            )
+            .FormActions(s => s.Self())
+            .FrameAncestors(s => s.Self())
+            .ImageSources(s => s.Self().CustomSources("data:"))
+            .DefaultSources(s => s.Self())
+            .ObjectSources(s => s.Self().CustomSources("data:"))
+            .FrameSources(s => s.Self().CustomSources("data:"))
+            .ScriptSources(s =>
+                s.Self()
+                    .CustomSources(
+                        "https://apis.google.com",
+                        "https://maxcdn.bootstrapcdn.com",
+                        "https://cdnjs.cloudflare.com",
+                        "https://code.jquery.com",
+                        "https://stackpath.bootstrapcdn.com",
+                        "https://fonts.googleapis.com"
+                    )
+            );
+    });
+}
+
+app.UseXXssProtection(options => options.EnabledWithBlockMode());
+app.UseNoCacheHttpHeaders();
+
+app.UseCookiePolicy(
+    new CookiePolicyOptions
+    {
+        HttpOnly = HttpOnlyPolicy.Always,
+        Secure = CookieSecurePolicy.Always,
+        MinimumSameSitePolicy = SameSiteMode.None,
+    }
+);
+
+app.UseHttpsRedirection();
+app.UseRouting();
+app.MapControllers();
+
+app.Run();
+
+// ---------------------------------------------------------------------------
+// Serilog configuration
+// ---------------------------------------------------------------------------
+static void ConfigureLogging(IHostEnvironment env, IConfiguration configuration)
+{
+    var loggerConfiguration = new LoggerConfiguration()
+        .Enrich.FromLogContext()
+        .Enrich.WithExceptionDetails()
+        .Enrich.WithMachineName()
+        .Enrich.WithProperty("app", "Restitution")
+        .Enrich.WithProperty("environment", env.EnvironmentName)
+        .Enrich.WithEnvironmentUserName()
+        .Enrich.WithCorrelationId()
+        .Enrich.WithSpan()
+        .Enrich.WithProperty(
+            "version",
+            Assembly.GetExecutingAssembly().GetName().Version?.ToString() ?? "Unknown"
+        )
+        .Enrich.WithProperty("UTC_Timestamp", DateTime.UtcNow.ToString("o"));
+
+    if (env.IsDevelopment())
+    {
+        loggerConfiguration.MinimumLevel.Debug();
+    }
+    else
+    {
+        loggerConfiguration.MinimumLevel.Information();
+    }
+
+    loggerConfiguration
+        .MinimumLevel.Override("Microsoft", Serilog.Events.LogEventLevel.Warning)
+        .MinimumLevel.Override("System", Serilog.Events.LogEventLevel.Warning);
+
+    loggerConfiguration.WriteTo.Console();
+
+    var splunkCollectorUrl = configuration["SPLUNK_COLLECTOR_URL"];
+    var splunkToken = configuration["SPLUNK_TOKEN"];
+
+    if (!string.IsNullOrEmpty(splunkCollectorUrl) && !string.IsNullOrEmpty(splunkToken))
+    {
+        HttpClientHandler? handler = null;
+
+        if (env.IsDevelopment())
+        {
+            handler = new HttpClientHandler
+            {
+                ServerCertificateCustomValidationCallback =
+                    HttpClientHandler.DangerousAcceptAnyServerCertificateValidator,
+            };
+        }
+
+        loggerConfiguration.WriteTo.EventCollector(
+            splunkHost: splunkCollectorUrl,
+            eventCollectorToken: splunkToken,
+            sourceType: "coast:restitution:api",
+            restrictedToMinimumLevel: Serilog.Events.LogEventLevel.Information,
+            messageHandler: handler,
+            batchSizeLimit: 100,
+            batchIntervalInSeconds: 2
+        );
+    }
+
+    Log.Logger = loggerConfiguration.CreateLogger();
+
+    Serilog.Debugging.SelfLog.Enable(msg =>
+    {
+        Console.Error.WriteLine($"Serilog Error: {msg}");
+    });
+
+    Log.Logger.Information("Restitution API Started");
 }
