+++ restitution-app/restitution-app.csproj
@@ -7,7 +7,6 @@
     <BuildServerSideRenderer>false</BuildServerSideRenderer>
     <RootNamespace>Gov.Cscp.VictimServices.Public</RootNamespace>
     <UserSecretsId>ea444fef-9537-4671-aafd-138a4dc8d981</UserSecretsId>
-    <StartupObject>Gov.Cscp.VictimServices.Public.Program</StartupObject>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Azure.Core" Version="1.51.1" />
